agent_system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
sqlmodel==0.0.14
pydantic==2.5.0
pydantic-settings==2.1.0
aiosqlite==0.19.0
asyncio-mqtt==0.16.1
cryptography==41.0.7
python-dotenv==1.0.0
authlib==1.2.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gitpython==3.1.40
anthropic==0.7.7
openai==1.3.7
google-generativeai==0.8.3
psutil==5.9.6
aiofiles==23.2.1
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
PyYAML==6.0.1

# Official MCP Python SDK
mcp>=1.0.0

# Additional MCP integrations (optional)
# fastmcp>=2.0.0  # Enhanced MCP framework
# shell-command-mcp  # Shell command execution
# github-mcp-server  # GitHub integration